% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continents.R
\docType{data}
\name{continents}
\alias{continents}
\title{Continent names for Worldclim pre-crop layers}
\format{
## `continents`
A data frame with 9 rows and 2 columns:
\describe{
  \item{id}{id of the continent name}
  \item{continent}{Valid name continent name}
}
}
\source{
R Natural Earth project: \url{https://ropensci.github.io/rnaturalearth/}
}
\usage{
continents
}
\description{
A list of valid continent names for Worldclim pre-crop layers
}
\keyword{datasets}
