% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envLayers.R
\docType{data}
\name{envLayers}
\alias{envLayers}
\title{Environmental bioclimatic layers from worldclim}
\format{
## `envLayers`
A data frame with 9 rows and 2 columns:
\describe{
  \item{name}{id of the environmental variable}
  \item{layer}{Envirionmental bioclimatic layer valid name}
}
}
\source{
WorldClim 2.0 project: \url{https://www.worldclim.org/}
}
\usage{
envLayers
}
\description{
A list of valid bioclimatic layer names from Worldclim
}
\keyword{datasets}
