% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bio_layer.R
\name{get_bio_layer}
\alias{get_bio_layer}
\title{Function to return Worldclim layer}
\source{
WorldClim 2.0 project: \url{https://www.worldclim.org/}
get_bio_layer(layer = "test")
}
\usage{
get_bio_layer(layer, continent = "world", dir = NULL)
}
\arguments{
\item{layer}{The bioclimatic layer want to download}

\item{continent}{The continent from which the bioclimatic layers are required}

\item{dir}{A directory where to write the output}
}
\value{
A SpatRaster object with world bioclimatic raster
}
\description{
Function to return Worldclim layer
}
