% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bioclim.R
\name{get_bioclim}
\alias{get_bioclim}
\title{Worldclim 2.0 bioclimatic layers by continent}
\source{
WorldClim 2.0 project: \url{https://www.worldclim.org/}
}
\usage{
get_bioclim(continent = "World")
}
\arguments{
\item{continent}{The continent from which the bioclimatic layers are required}
}
\value{
A SpatVector object with bioclimatic layers by continent
}
\description{
Worldclim 2.0 bioclimatic layers by continent
}
\examples{
get_bioclim(continent = "test")
}
